rename effect.async to effect.parallel_async, replace async.py with 'from .parallel_async import *'
this should hopefully allow us to be backwards compatible so older versions of python can still use effect.async, but to switch to python 3.7 you will need to import effect.parallel_async.
