+++ effect/test_async.py
@@ -5,7 +5,7 @@ from testtools.testcase import TestCase
 from ._base import Effect, perform
 from ._dispatcher import ComposedDispatcher, TypeDispatcher
 from ._intents import ParallelEffects, base_dispatcher, parallel
-from .async import perform_parallel_async
+from .parallel_async import perform_parallel_async
 from .test_base import func_dispatcher
 from .test_parallel_performers import ParallelPerformerTestsMixin
